% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{c_bounds}
\alias{c_bounds}
\title{Calculate C set of bounds}
\usage{
c_bounds(
  b_lower,
  b_upper,
  rho_upper,
  rho_lower = 1/rho_upper,
  return_full_set = FALSE
)
}
\arguments{
\item{b_lower}{The lower bound on \eqn{\nu} in the second setting (B in the paper).}

\item{b_upper}{The upper bound on \eqn{\nu} in the second setting (B in the paper).}

\item{rho_upper}{The upper bound on the difference of biases between the two settings.}

\item{rho_lower}{The lower bound on the difference of biases between the two settings. By default, this is set to \code{1/rho_upper}, which is a common choice that ensures a symmetrical bound on the ratio.}

\item{return_full_set}{A boolean indicating whether to return the full set of bounds (the four values corresponding to the corners of the rectangle defined by the bounds) or just the range (the minimum and maximum of those values). By default, this is set to \code{FALSE}, which returns just the range.}
}
\description{
Calculate the set that bounds the difference between two effects, from Huntington-Klein (2025).
}
