% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_estimate.R
\name{estimate_het_effects}
\alias{estimate_het_effects}
\title{Estimate Effects by Setting}
\usage{
estimate_het_effects(
  data,
  setting,
  treatment,
  outcome = NULL,
  covariates = NULL,
  formula = NULL,
  estimation_function = hetset_feols,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{setting}{A string or string vector specifying the name of a variable or variables that define the different estimation settings.}

\item{treatment}{A string specifying the name of the treatment variable.}

\item{outcome}{A string specifying the name of the outcome variable.}

\item{covariates}{A string vector specifying the names of covariate variables to include in the estimation.}

\item{formula}{An optional formula object specifying the model to be estimated. Should follow whatever formula syntax is allowed by \code{estimation_function}. If provided, this overrides the \code{outcome} and \code{covariates} arguments. The treatment variable should still be specified via the \code{treatment} argument.}

\item{estimation_function}{A function that takes the above-mentioned arguments and returns a data frame of estimates by setting, in the format outlined in \code{estimation_wrapper}.}

\item{verbose}{Logical indicating whether to print progress messages. Default is \code{FALSE}.}

\item{...}{Additional arguments to pass to \code{estimation_function}.}
}
\value{
A data frame containing the estimated effects by setting, along with estimation statistics like sample size, standard errors, and model fit.
}
\description{
Estimate the effects of a treatment variable on an outcome variable within different settings defined by a set of covariates.
}
\details{
This function works with a wide range of estimation functions. By default it is set up to work with linear regression via \code{fixest::feols}, specifically via the wrapper \code{hetset_feols}.

See the function \code{estimation_wrapper} to create custom estimation functions compatible with this function.
}
