% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_relationship_bounds.R
\name{supershort_rho_bounds_proposal}
\alias{supershort_rho_bounds_proposal}
\title{Propose a set of rho bounds}
\usage{
supershort_rho_bounds_proposal(
  data,
  setting,
  treatment,
  model,
  model_shorter,
  treatment_code = "coef(m)[treatment]",
  negative_is_uninformative = TRUE,
  minimum_bound = NULL,
  maximum_bound = NULL,
  settings_order = NULL,
  return = "matrix"
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{setting}{A string or string vector specifying the name of a variable or variables that define the different estimation settings.}

\item{treatment}{A string specifying the name of the treatment variable.}

\item{model}{A string specifying the model to be estimated including the estimation function. Should refer to the name of the data set as \code{data}. For example, \code{'fixest::feols(formula = y ~ x + a + b, data = data)'}. This model should include all covariates, including those that are to be dropped.}

\item{model_shorter}{A string specifying the model to be estimated including the estimation function. Should refer to the name of the data set as \code{data}. For example, \code{'fixest::feols(formula = y ~ x + a, data = data)'}. This model should not include the predictors further removed to create a super-short model.}

\item{treatment_code}{A string specifying how to extract the treatment effect from a model named \code{m}.}

\item{negative_is_uninformative}{If the change in bias in setting A and setting B are of opposite signs, consider this analysis to be uninformative.}

\item{minimum_bound}{A number greater than 1. All rho bounds between 1 and \code{minimum_bound} will be set to \code{minimum_bound}, and all rho bounds between 1 and \code{1/minimum_bound} will be set to \code{1/minimum_bound}.}

\item{maximum_bound}{A number greater than 1. All rho bounds will be constrained to \code{[1/maximum_bound,maximum_bound]}. Uninformative bounds given by \code{negative_is_uninformative = TRUE} will not be adjusted.}

\item{settings_order}{A vector with the values of the settings in the order they are used in \code{rho_bounds}. Defaults to \code{sort()} order.}

\item{return}{Set to \code{'table'} to return a table of rho bounds, \code{'matrix'} to return a bounds matrix suitable for use with \code{build_bounds}, or \code{'both'} to get a list with both.}
}
\description{
This function suggests a plausible bound for rho, the proportional parameter relating the omitted variable bias in setting A with the omitted variable bias in setting B.
}
\details{
This bound is based on taking the estimation model (which is already missing some predictors, since we have omitted variable bias) and omitting more variables. The change in the estimated coefficient is the omitted variable bias added by the further removal of predictors. The ratio of the change in biases in the two settings is given as their proportional relationship.
}
\examples{
library(fixest)
library(hetset)
data("close_college")

bounds = ovb_bounds_by_setting(close_college,
                               setting = 'blacksouth',
                               treatment = 'educ',
                               model = 'feols(lwage ~ educ + smsa + married + exper, data = data)',
                               benchmark_covariates = c('smsa','married','exper'),
                               kd = .1)
boundmat = supershort_rho_bounds_proposal(
  data = close_college,
  setting = 'blacksouth',
  treatment = 'educ',
  model = 'fixest::feols(lwage ~ educ + smsa + married + exper, data = data)',
  model_shorter = 'fixest::feols(lwage ~ educ, data = data)',
  settings_order = c('Neither','Black','South','Black and South'))

# Get the partial identification bounds
pi_bounds = build_bounds(bounds, boundmat,
                            settings_order = c('Neither','Black','South','Black and South'))

# Get the univariate bounds table
univariate_bounds_table(pi_bounds)

}
