% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sets.R
\name{build_bounds}
\alias{build_bounds}
\title{Build the full set of bounds for effect pairs}
\usage{
build_bounds(
  bounds_data,
  rho_bounds,
  settings_order = bounds_data$setting,
  ...
)
}
\arguments{
\item{bounds_data}{A data frame containing the partial identification bounds for each individual setting's effect, from a function like \code{ovb_bounds_by_setting}. Must be a \code{data.frame} with a column for \code{setting} and columns for \code{original_estimate}, \code{adjusted_estimate}, \code{lower_plausible_bound}, and \code{upper_plausible_bound}. Other columns are acceptable.}

\item{rho_bounds}{A matrix with the upper bound of rho_ij in the ith row and jth column, and the lower bound in the ith column and jth row. Trace values will be ignored. If all lower triangle values are 0, they will be set to the inverse of the corresponding upper bound. Alternately, provide a single real value to bound all pairs at that same value and its inverse. Set a bound to 0 to ignore it (set no restriction).}

\item{settings_order}{A vector with the values of the settings in the order they are used in \code{rho_bounds}. Defaults to the order of settings in \code{bounds_data}.}

\item{...}{Additional arguments (not currently used).}
}
\description{
Build the full set of bounds for effect pairs based on partial identification bounds.
}
\examples{

library(fixest)
library(hetset)
data("close_college")

bounds = ovb_bounds_by_setting(close_college,
                               setting = 'blacksouth',
                               treatment = 'educ',
                               model = 'feols(lwage ~ educ + smsa + married + exper, data = data)',
                               benchmark_covariates = c('smsa','married','exper'),
                               kd = .1)

# Put a 1.5 bound on Black vs Black & South and Neither vs South,
# 2 on Neither vs. black and South vs. Black & South
# and no restriction on Neither vs Black & South and on Black vs. South (setting to 0)
# Setting all the lower triangle values of 0
# will make them automatically the inverse of the upper triangle values
boundmat = matrix(c(0, 2, 1.5, 0,
                    0, 0, 0, 1.5,
                    0, 0, 0, 2,
                    0, 0, 0,0), nrow = 4, byrow = TRUE)
# Get the partial identification bounds
pi_bounds = build_bounds(bounds, boundmat,
                            settings_order = c('Neither','Black','South','Black and South'))

}
