% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sets.R
\name{identified_set_exists}
\alias{identified_set_exists}
\title{Check for the existence of an identified set satisfying all bounds}
\usage{
identified_set_exists(
  full_bounds_set,
  check_only = NULL,
  return_one_solution = FALSE
)
}
\arguments{
\item{full_bounds_set}{A full set of univariate and paired bounds produced by \code{build_bounds}}

\item{check_only}{A vector of setting names to check only those settings. If \code{NULL} (default) checks all settings. This option is used to try to figure out which setting combinations might be leading to no identified set existing.}

\item{return_one_solution}{Set to \code{TRUE} to return a set of parameters that satisfy all bounds. \emph{IT IS EXTREMELY IMPORTANT TO NOTE} that this is not a "best" solution or anything like that. It is simply one possible solution.}
}
\description{
Check for the existence of an identified set satisfying all bounds
}
\examples{
library(fixest)
library(hetset)
data("close_college")

bounds = ovb_bounds_by_setting(close_college,
                               setting = 'blacksouth',
                               treatment = 'educ',
                               model = 'feols(lwage ~ educ + smsa + married + exper, data = data)',
                               benchmark_covariates = c('smsa','married','exper'),
                               kd = .1)

# Put a 1.5 bound on Black vs Black & South and Neither vs South,
# 2 on Neither vs. black and South vs. Black & South
# and no restriction on Neither vs Black & South and on Black vs. South (setting to 0)
# Setting all the lower triangle values of 0
# will make them automatically the inverse of the upper triangle values
boundmat = matrix(c(0, 2, 1.5, 0,
                    0, 0, 0, 1.5,
                    0, 0, 0, 2,
                    0, 0, 0,0), nrow = 4, byrow = TRUE)
# Get the partial identification bounds
boundset = build_bounds(bounds, boundmat,
                        settings_order = c('Neither','Black','South','Black and South'))


# Check if there exists a set of effect parameters that satisfy all bounds.
identified_set_exists(boundset)
# Get one example of a set of parameters in the identified set.
identified_set_exists(boundset, return_one_solution = TRUE)

}
