% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_estimate_sensemakr.R
\name{ovb_bounds_by_setting}
\alias{ovb_bounds_by_setting}
\title{Estimate Partial Identification Effects by Setting with sensemakr}
\usage{
ovb_bounds_by_setting(
  data,
  setting,
  treatment,
  model,
  benchmark_covariates,
  kd = 1,
  ky = kd,
  estimation_function = fixest::feols,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{setting}{A string or string vector specifying the name of a variable or variables that define the different estimation settings.}

\item{treatment}{A string specifying the name of the treatment variable.}

\item{model}{A string specifying the model to be estimated including the estimation function. Should refer to the name of the data set as \code{data}. For example, \code{'fixest::feols(formula = y ~ x + a + b, data = data)'}.}

\item{benchmark_covariates}{A string vector specifying the names of baseline covariates to be used for partial identification with \code{sensemakr::ovb_bounds}.}

\item{kd}{A numeric vector. Parameterizes how many times stronger the confounder is related to the treatment in comparison to the observed benchmark covariate. Default value is 1 (confounder is as strong as benchmark covariate).}

\item{ky}{A numeric vector. Parameterizes how many times stronger the confounder is related to the outcome in comparison to the observed benchmark covariate. Default value is the same as \code{kd}.}

\item{estimation_function}{The function that will be used to estimate the effect.}

\item{verbose}{Logical indicating whether to print progress messages. Default is \code{FALSE}.}

\item{...}{Additional arguments to pass to \code{sensemakr::ovb_bounds}.}
}
\value{
A data frame containing the estimated effects by setting, along with estimation statistics like sample size, standard errors, and model fit.
}
\description{
Estimate the effects of a treatment variable on an outcome variable within different settings defined by a set of covariates, with partial identification bounds based on a set of baseline covariates with \code{sensemakr::ovb_bounds}.
}
\examples{

library(fixest)
library(hetset)
data("close_college")

bounds = ovb_bounds_by_setting(close_college,
                               setting = 'blacksouth',
                               treatment = 'educ',
                               model = 'feols(lwage ~ educ + smsa + married + exper, data = data)',
                               benchmark_covariates = c('smsa','married','exper'),
                               kd = .1)

}
