% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sets.R
\name{univariate_bounds_table}
\alias{univariate_bounds_table}
\title{Univariate Bounds Table}
\usage{
univariate_bounds_table(full_bounds_set, pin_values = NULL, ...)
}
\arguments{
\item{full_bounds_set}{A full set of univariate and paired bounds produced by \code{build_bounds}.}

\item{pin_values}{A named list of settings and values to pin estimates in those settings to. Values can be numeric, or can be \code{'lower'}, \code{'upper'}, or \code{'original'} for the univariate lower-bound, upper bound, or original estimate, respectively. Useful for exploring the multivariate nature of the bounds.}

\item{...}{Additional arguments (not currently used).}
}
\description{
Creates a table of bounds for the effect in each setting.
}
\details{
The table shows the widest possible bounds consistent with the assumptions for each setting one at a time.

Note that because this is a highly multidimensional problem, the actual identified set of plausible estimates is somewhat narrower.

For example, if in setting A the effect must be between 0 and 1, in setting B it must be between 1 and 2, and the difference between them must be no greater than .5, then the univariate set for A will be .5 to 1 and the set for B will be 1 to 1.5, ignoring that the joint set (.5 for A, 1.5 for B) is not in the identified set.

To check whether a specific combination of values is in the identified set, use the function \code{valid_solution_check}.
}
\examples{
library(fixest)
library(hetset)
data("close_college")
bounds = ovb_bounds_by_setting(close_college,
                               setting = 'blacksouth',
                               treatment = 'educ',
                               model = 'feols(lwage ~ educ + smsa + married + exper, data = data)',
                               benchmark_covariates = c('smsa','married','exper'),
                               kd = .1)

# Put a 1.5 bound on Black vs Black & South and Neither vs South,
# 2 on Neither vs. black and South vs. Black & South
# and no restriction on Neither vs Black & South and on Black vs. South (setting to 0)
# Setting all the lower triangle values of 0 will make them automatically the inverse of the upper triangle values
boundmat = matrix(c(0, 2, 1.5, 0,
                    0, 0, 0, 1.5,
                    0, 0, 0, 2,
                    0, 0, 0,0), nrow = 4, byrow = TRUE)
# Get the partial identification bounds
boundset = build_bounds(bounds, boundmat,
                        settings_order = c('Neither','Black','South','Black and South'))

# Get the univariate bounds table
univariate_bounds_table(boundset)
# Note in this example that the bounds are only very slightly tighter than if we hadn't worried about cross-setting relationships (for Black).

}
